paymentMethods-BugDelete
pulling an element of an array of Object ID is not quite right yet

+++ src/app/controller/accountBank.controller.ts
@@ -109,6 +109,11 @@ export class AccountBankController {
 
       const user = await userService.findUser(userId, producerModels);
 
+      if(!user){
+        Logger.errorLog("Usuario não encontrado")
+        return badRequestResponse(res)
+      }
+
       await ReceiptAccountService.deleteItems(id)
         .then(() => {
           Logger.infoLog("Delete id from user model");
